Rename Smokey to Leash!
