+++ Gemfile
@@ -2,5 +2,5 @@
 
 source "https://rubygems.org"
 
-# Specify your gem's dependencies in smokey_gem.gemspec
+# Specify your gem's dependencies in leash.gemspec
 gemspec
