+++ exe/leash
@@ -1,6 +1,6 @@
 #! /usr/bin/env ruby
 # frozen_string_literal: true
 
-require "smokey_gem"
+require "leash"
 
-SmokeyGem::CLI.start(ARGV)
+Leash::CLI.start(ARGV)
