+++ lib/leash.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 # :nodoc:
-module SmokeyGem
+module Leash
   def self.root_join(*joins)
     path_parts = [File.dirname(__FILE__)] + joins.map(&:to_s)
     joined_path = File.join(*path_parts)
@@ -9,8 +9,8 @@ module SmokeyGem
   end
 end
 
-require_relative "smokey_gem/version"
-require_relative "smokey_gem/cli"
-require_relative "smokey_gem/templates"
-require_relative "smokey_gem/configuration"
-require_relative "smokey_gem/docker_compose"
+require_relative "leash/version"
+require_relative "leash/cli"
+require_relative "leash/templates"
+require_relative "leash/configuration"
+require_relative "leash/docker_compose"
