+++ lib/leash/docker_compose.rb
