+++ lib/leash/docker_compose/service.rb
