+++ templates/smokey_config.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-SmokeyGem::Configuration.new do |config|
+Leash::Configuration.new do |config|
   # Do stuff here.
 
   # Example dockerfile config.
